+++ src/stream/src/executor/source/fetch_executor.rs
@@ -33,6 +33,7 @@ use risingwave_connector::ConnectorParams;
 use risingwave_source::source_desc::SourceDesc;
 use risingwave_storage::store::PrefetchOptions;
 use risingwave_storage::StateStore;
+use thiserror_ext::AsReport;
 
 use crate::executor::stream_reader::StreamReaderWithPause;
 use crate::executor::{
@@ -203,7 +204,7 @@ impl<S: StateStore> FsFetchExecutor<S> {
         while let Some(msg) = stream.next().await {
             match msg {
                 Err(e) => {
-                    tracing::error!("Fetch Error: {:?}", e);
+                    tracing::error!(error = %e.as_report(), "Fetch Error");
                     splits_on_fetch = 0;
                 }
                 Ok(msg) => {